+++ integration/example_test.go
@@ -1,3 +1,5 @@
+//go:build openssl
+
 package integration_test
 
 import (